requirements.txt
config
praw
numpy
scikit-learn
matplotlib
seaborn
python-dotenv